fix fabric conversion issues

+++ common/src/main/java/com/fej1fun/potentials/fluid/BaseFluidStorage.java
@@ -95,7 +95,7 @@ public class BaseFluidStorage implements UniversalFluidStorage {
     }
 
     @Override
-    public FluidStack drain(int maxAmount, boolean simulate) {
+    public FluidStack drain(long maxAmount, boolean simulate) {
         AtomicReference<FluidStack> toReturn = new AtomicReference<>(FluidStack.empty());
         fluidStacks.stream().filter(stack -> !stack.isEmpty()).max(Comparator.comparing(FluidStack::getAmount)).ifPresent(stack -> {
             long removedAmount = Math.min(maxAmount, stack.getAmount());
